+++ R/greenAlgo-package.R
@@ -1,0 +1,7 @@
+#' \code{greenAlgoR} package
+#'
+#'  This package computes ecological footprint in R (based on [green-algorithms](https://calculator.green-algorithms.org/).
+#'  greenAlgoR also made it simple to compute ecological footprint of {[targets](https://github.com/ropensci/targets)} pipelines..
+#' @name greenAlgoR-package
+#' @import ggplot2 targets benchmarkme
+NULL