+++ NEWS.md
@@ -1,3 +1,4 @@
 # greenAlgoR 0.1
 
-* Initial CRAN submission.
+* Initial github submission.
+* Creation of function `ga_footprint()` and `ga_targets()`